requirements.txt
aiofiles==25.1.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
evdev==1.9.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==25.0
pillow==12.0.0
pynput==1.8.1
pytesseract==0.3.13
python-telegram-bot==22.5
python-xlib==0.33
requests==2.32.5
six==1.17.0
urllib3==2.6.2
